requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.20.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
pyspark==3.5.1